separated responsibilities collecting and processing
